pyproject.toml
[tool.poetry]
name = "gitstats"
version = "0.1.0"
description = "A Python CLI tool to analyze Git repository statistics for developers"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.40"
tabulate = "^0.9.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gitstats = "gitstats:main" 